+++ packages/toolpad-app/src/toolpadDataSources/function/runtime/text-encoding.ts
@@ -1,0 +1,3 @@
+import 'fastestsmallesttextencoderdecoder';
+
+export {};
