+++ packages/toolpad-app/src/toolpadDataSources/function/runtime/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../../../tsconfig.json",
+  "compilerOptions": {
+    "lib": ["es6"]
+  },
+  "include": ["./**/*.ts"]
+}
