+++ packages/toolpad-app/src/toolpadDataSources/function/runtime/url.ts
@@ -1,0 +1,6 @@
+import { URL, URLSearchParams } from 'whatwg-url';
+
+global.URL = URL;
+global.URLSearchParams = URLSearchParams;
+
+export {};
